+++ src/app/app.rs
@@ -170,11 +170,16 @@ impl App {
         skin: &AppSkin,
         con: &AppContext,
     ) -> Result<(), ProgramError> {
+        // if some images are displayed by kitty, we'll erase it,
+        // but only after having displayed the new ones (if any)
+        // to prevent some flickerings
         #[cfg(unix)]
-        if let Some(renderer) = kitty::image_renderer() {
-            let mut renderer = renderer.lock().unwrap();
-            renderer.erase_images(w)?;
-        }
+        let previous_images = kitty::image_renderer()
+            .as_ref()
+            .and_then(|renderer| {
+                let mut renderer = renderer.lock().unwrap();
+                renderer.take_current_images()
+            });
         for (idx, panel) in self.panels.as_mut_slice().iter_mut().enumerate() {
             let focused = idx == self.active_panel_idx;
             let skin = if focused { &skin.focused } else { &skin.unfocused };
@@ -184,6 +189,14 @@ impl App {
                 panel.display(w, focused, self.screen, skin, con)?,
             );
         }
+        #[cfg(unix)]
+        if let Some(previous_images) = previous_images {
+            if let Some(renderer) = kitty::image_renderer().as_ref() {
+                let mut renderer = renderer.lock().unwrap();
+                renderer.erase(w, previous_images)?;
+            }
+        }
+        w.flush()?;
         Ok(())
     }
 
@@ -501,22 +514,28 @@ impl App {
             ))
             .transpose()?;
 
+        let mut skip_redraw = false;
         loop {
-            if !self.quitting {
+            if !self.quitting && !skip_redraw {
                 self.display_panels(w, &skin, con)?;
-                w.flush()?;
                 if self.do_pending_tasks(con, &mut dam)? {
                     let other_path = self.get_other_panel_path();
                     self.mut_panel().refresh_input_status(&other_path, con);
                     self.display_panels(w, &skin, con)?;
-                    w.flush()?;
                 }
             }
 
+            skip_redraw = false;
             match dam.next(&self.rx_seqs) {
                 Either::First(Some(event)) => {
-                    debug!("event: {:?}", &event);
+                    info!("event: {:?}", &event);
                     match event {
+                        Event::EscapeSequence(escape_sequence) => {
+                            debug!("received escape sequence {}", escape_sequence);
+                            // this escape sequence may be the result of a drawing
+                            // operation so we must avoid redrawing now
+                            skip_redraw = true;
+                        }
                         Event::Click(x, y, KeyModifiers::NONE)
                             if self.clicked_panel_index(x, y) != self.active_panel_idx =>
                         {
@@ -547,7 +566,7 @@ impl App {
                     break;
                 }
                 Either::Second(Some(raw_sequence)) => {
-                    debug!("got sequence: {:?}", &raw_sequence);
+                    debug!("got command sequence: {:?}", &raw_sequence);
                     for (input, arg_cmd) in raw_sequence.parse(con)? {
                         self.mut_panel().set_input_content(&input);
                         self.apply_command(w, arg_cmd, &skin.focused, con)?;
@@ -556,10 +575,8 @@ impl App {
                             return Ok(self.launch_at_end.take());
                         } else {
                             self.display_panels(w, &skin, con)?;
-                            w.flush()?;
                             if self.do_pending_tasks(con, &mut dam)? {
                                 self.display_panels(w, &skin, con)?;
-                                w.flush()?;
                             }
                         }
                     }