+++ src/kitty/image_renderer.rs
@@ -14,10 +14,15 @@ use {
     std::{
         env,
         io::{self, Write},
-        path::Path,
     },
     tempfile,
 };
+
+pub type KittyImageSet = Vec<usize>;
+
+/// The max size of a data payload in a kitty escape sequence
+/// according to kitty's documentation
+const CHUNK_SIZE: usize = 4096;
 
 /// until I'm told there's another terminal supporting the kitty
 /// terminal, I think I can just check the name
@@ -34,10 +39,22 @@ fn div_ceil(a: u32, b: u32) -> u32 {
     a / b + (0 != a % b) as u32
 }
 
+/// the image renderer, with knowledge of the
+/// console cells dimensions, and built only on Kitty.
+///
+/// There are two ways to display the image:
+/// - print_with_temp_file, which is faster on my
+///    computers (with SSD)
+/// - print_with_chunks which looks slightly less
+///    racy and might be faster on computers with
+///    very slow disks
+/// Note that I didn't test yet the named shared memory
+/// solution offered by kitty.
 pub struct KittyImageRenderer {
     cell_width: u32,
     cell_height: u32,
-    has_image_on_screen: bool,
+    next_id: usize,
+    current_images: Option<KittyImageSet>,
 }
 
 impl KittyImageRenderer {
@@ -50,8 +67,12 @@ impl KittyImageRenderer {
             .map(|(cell_width, cell_height)| Self {
                 cell_width,
                 cell_height,
-                has_image_on_screen: false,
+                current_images: None,
+                next_id: 1,
             })
+    }
+    pub fn take_current_images(&mut self) -> Option<KittyImageSet> {
+        self.current_images.take()
     }
     fn rendering_dim(
         &self,
@@ -77,6 +98,17 @@ impl KittyImageRenderer {
             (optimal_cols * area_rows / optimal_rows, area_rows)
         }
     }
+    /// return a new image id which is assumed will be used
+    fn new_id(&mut self) -> usize {
+        let new_id = self.next_id;
+        self.next_id += 1;
+        self.current_images
+            .get_or_insert_with(Vec::new)
+            .push(new_id);
+        new_id
+    }
+    /// render the image by sending multiple kitty escape sequence, each
+    /// one with part of the image raw data (encoded as base64)
     pub fn print_with_chunks(
         &mut self,
         w: &mut W,
@@ -89,19 +121,21 @@ impl KittyImageRenderer {
         let bytes = rgba.as_raw();
         let encoded = base64::encode(bytes);
         let (c, r) = self.rendering_dim(width, height, cols.into(), rows.into());
+        let id = self.new_id();
         debug!("rendering_dim: {:?}", (c, r));
         let mut pos = 0;
         loop {
-            if pos + 4096 < encoded.len() {
+            if pos + CHUNK_SIZE < encoded.len() {
                 write!(w,
-                    "\u{1b}_Ga=T,f=32,t=d,s={},v={},c={},r={},m=1;{}\u{1b}\\",
+                    "\u{1b}_Ga=T,f=32,t=d,i={},s={},v={},c={},r={},m=1;{}\u{1b}\\",
+                    id,
                     width,
                     height,
                     c,
                     r,
-                    &encoded[pos..pos+4096],
+                    &encoded[pos..pos+CHUNK_SIZE],
                 )?;
-                pos += 4096;
+                pos += CHUNK_SIZE;
             } else {
                 // last chunk
                 write!(w,
@@ -111,33 +145,11 @@ impl KittyImageRenderer {
                 break;
             }
         }
-        self.has_image_on_screen = true;
         Ok(())
     }
-    // deprecated
-    pub fn print_png(
-        &mut self,
-        w: &mut W,
-        path: &Path,
-        cols: u16,
-        rows: u16,
-    ) -> Result<(), ProgramError> {
-        let path = path.to_str()
-            .ok_or_else(|| io::Error::new(
-                io::ErrorKind::Other,
-                "Path can't be converted to UTF8",
-            ))?;
-        let encoded_path = base64::encode(path);
-        debug!("printing png with kitty c={}, r={}, path={}", cols, rows, &path);
-        write!(w,
-            "\u{1b}_Ga=T,f=100,t=f,r={},c={};{}\u{1b}\\",
-            rows,
-            cols,
-            encoded_path,
-        )?;
-        self.has_image_on_screen = true;
-        Ok(())
-    }
+    /// render the image by writing the raw data in a temporary file
+    /// then giving to kitty the path to this file in the payload of
+    /// a unique kitty ecape sequence
     pub fn print_with_temp_file(
         &mut self,
         w: &mut W,
@@ -166,8 +178,10 @@ impl KittyImageRenderer {
         let encoded_path = base64::encode(path);
         debug!("temp file written: {:?}", path);
         let (c, r) = self.rendering_dim(width, height, cols.into(), rows.into());
+        let id = self.new_id();
         write!(w,
-            "\u{1b}_Ga=T,f=32,t=t,s={},v={},c={},r={};{}\u{1b}\\",
+            "\u{1b}_Ga=T,f=32,t=t,i={},s={},v={},c={},r={};{}\u{1b}\\",
+            id,
             width,
             height,
             c,
@@ -175,17 +189,26 @@ impl KittyImageRenderer {
             encoded_path,
         )?;
         debug!("file len: {}", temp_file.metadata().unwrap().len());
-        self.has_image_on_screen = true;
         Ok(())
     }
-    pub fn erase_images(
+    pub fn erase(
+        &mut self,
+        w: &mut W,
+        ids: KittyImageSet,
+    ) -> Result<(), ProgramError> {
+        for id in ids {
+            debug!("erase kitty image {}", id);
+            write!(w, "\u{1b}_Ga=d,d=I,i={}\u{1b}\\", id)?;
+        }
+        Ok(())
+    }
+    /// erase all kitty images, even the forgetted ones
+    pub fn erase_all(
         &mut self,
         w: &mut W,
     ) -> Result<(), ProgramError> {
-        if self.has_image_on_screen {
-            write!(w, "\u{1b}_Ga=d\u{1b}\\")?;
-            self.has_image_on_screen = false;
-        }
+        write!(w, "\u{1b}_Ga=d,d=A\u{1b}\\")?;
+        self.current_images = None;
         Ok(())
     }
 }